requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
openai
pandas
requests
numpy
faiss-cpu
scikit-learn
